Reload workspaces when starting a search
Loading workspaces from sqlite is so fast we can just do it whenever we
start a search.  We only do it at the beginning of a search to make
sure that item IDs do not change while the user refines their search.

Remove the reload interface which no longer serves any purpose now.

Do not load workspaces at startup, because we'd throw away those
workspaces when reloading workspaces on the first search.

Simplify reloading by taking a direct &mut reference to the storage
which is perfectly save as we're calling it from a &mut self context.