Parse desktop entry with configparser
We do not need the full complexity of desktop entries with locales and
all; all we need is to get the icon out of it.  configparser does the
job, and is fully audited already, so we can get rid of the unaudited
freedesktop-desktop-entry dependency and its also unaudited transitive
dependencies.